+++ src/action-result/options.ts
@@ -1,4 +1,0 @@
-export interface ActionResultOptions<T> {
-	errorThreshold?: number;
-	payload?: T;
-}