+++ src/aliases.ts
@@ -1,0 +1,6 @@
+export type ToStringable = {toString: () => string};
+
+/* eslint-disable @typescript-eslint/no-explicit-any */
+export type ANY = any;
+export type json = any;
+/* eslint-enable @typescript-eslint/no-explicit-any */