+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.0.0]
+	Renamed package to Fate.
+	Fate can be serialized into a string and be instantiate from a serialized string.
+
 ## [v0.3.0]
 	Defaults to fail on 1 or more errors.
 	can not succeed if payload is not set.